+++ src/model/tools/draw_rectangle_imp.rs
@@ -122,8 +122,8 @@ impl Tool for DrawRectangleTool {
             .get_attribute()
             .get_foreground();
         for rect in lines.outline() {
-            for y in 0..rect.size.height {
-                for x in 0..rect.size.width {
+            for y in 0..(rect.size.height as i32) {
+                for x in 0..(rect.size.width as i32) {
                     set_half_block(
                         editor,
                         Position::new(rect.start.x + x, rect.start.y + y),
