+++ src/model/tools/pencil_imp.rs
@@ -46,8 +46,8 @@ impl PencilTool {
                         .caret
                         .get_attribute()
                         .get_foreground();
-                    for y in 0..rect.size.height {
-                        for x in 0..rect.size.width {
+                    for y in 0..(rect.size.height as i32) {
+                        for x in 0..(rect.size.width as i32) {
                             set_half_block(
                                 editor,
                                 Position::new(rect.start.x + x, rect.start.y + y),
@@ -220,7 +220,7 @@ pub fn draw_glyph_plain(editor: &AnsiEditor, ch: char, font_page: usize) -> impl
             if let Some(glyph) = font.get_glyph(ch) {
                 for y in 0..s.height {
                     for x in 0..s.width {
-                        if glyph.data[y as usize] & (128 >> x) != 0 {
+                        if glyph.data[y] & (128 >> x) != 0 {
                             painter.rect_filled(
                                 Rect::from_min_size(
                                     Pos2::new(
