+++ src/ui/dialogs/set_canvas_size_dialog.rs
@@ -7,8 +7,8 @@ use crate::{AnsiEditor, ModalDialog, TerminalResult};
 
 pub struct SetCanvasSizeDialog {
     pub should_commit: bool,
-    pub width: i32,
-    pub height: i32,
+    pub width: usize,
+    pub height: usize,
 }
 
 impl SetCanvasSizeDialog {
